feat: Wait for devices to be initialized
It's tricky to know whether a device is initialized due to the daemonless
and zero-config design which differ from other udev alternatives, so udev-zero
reports all devices with an initialized state.

This behaviour could be problematic especially when working with time-consuming
initializers and programs that heavily depend on udev with the "initialized"
filter e.g. NetworkManager.

Some files must be stored to indicate the state of a device if we want to keep
the daemonless design, so here "/dev/udevzero/{subsystem}/{sysname}" is chosen
e.g. "/dev/udevzero/net/lo", "/dev/udevzero/block/mmclk0". If the directory
exists, udev-zero will not set the initialized prop to 1 unless the
corresponding file exist; otherwise it assumes any device under the subsystem
has been initialized already.

The file may contain lines with the format of "{KEY}={VALUE}" to set extra
properties, e.g. the "SOUND_INITIALIZED" property being used in pulseaudio.